+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/title.js
@@ -59,10 +59,6 @@ const StyledDayCounter = styled.div`
   margin-left: calc(2 * ${COMPONENT_GUTTER});
 `
 
-const StyledStateDay = styled(Text)`
-  display: flex;
-`
-
 class TitleComponent extends PureComponent {
   static propTypes = {
     title: PropTypes.string,
@@ -70,7 +66,6 @@ class TitleComponent extends PureComponent {
     loading: PropTypes.bool,
     widgetTitle: PropTypes.string,
     dayCounter: PropTypes.element,
-    stateDay: PropTypes.string,
     titleProps: PropTypes.shape(),
   }
 
@@ -80,7 +75,6 @@ class TitleComponent extends PureComponent {
     loading: false,
     widgetTitle: '',
     dayCounter: null,
-    stateDay: '',
     titleProps: {},
   }
 
@@ -92,7 +86,6 @@ class TitleComponent extends PureComponent {
       loading,
       widgetTitle,
       dayCounter,
-      stateDay,
     } = this.props
 
     return (
@@ -141,13 +134,7 @@ class TitleComponent extends PureComponent {
   }
 
   renderWidgetTitle() {
-    const {
-      widgetTitle,
-      dayCounter,
-      stateDay,
-      loading,
-      titleTruncate,
-    } = this.props
+    const { widgetTitle, dayCounter, loading, titleTruncate } = this.props
 
     return (
       <>
@@ -171,16 +158,13 @@ class TitleComponent extends PureComponent {
 
           {!loading && (
             <StyledDayCounter>
-              <Text size="tiny" color="font1" weight="bold" lineHeight="normal">
+              <Text
+                size="tiny"
+                color="font1"
+                weight="regular"
+                lineHeight="normal"
+              >
                 {dayCounter}
-                <StyledStateDay
-                  size="micro"
-                  color="font1"
-                  weight="light"
-                  lineHeight="normal"
-                >
-                  {stateDay}
-                </StyledStateDay>
               </Text>
             </StyledDayCounter>
           )}