+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `CrowdfundingCard` mobile version.
+- Feature: Delete `stateDay` and update `dayCounter` props on `CrowdfundingCardWidget`
 
 ## [2.38.0] - 2019-11-08
 
@@ -19,7 +20,6 @@ Features:
 Fix:
 - Update `dayCounter` propTypes on `CrowdfundingCard`.
 
->>>>>>> origin
 ## [2.37.1] - 2019-11-04
 
 Fix: